requirements/base.txt
# Bleeding edge Django
django==1.8

# Configuration
django-environ==0.3.0
django-secure==1.0.1

# Forms
django-braces==1.8.0
django-crispy-forms==1.4.0
django-floppyforms==1.4.0

# Models
django-model-utils==2.2

# Images
Pillow==2.8.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.19.1

# For the persistence stores
psycopg2==2.6

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.7.2

# Time zones support
pytz==2015.2

# Your custom requirements go here
django-bootstrap3==5.4.0
django-ipware==1.1.0
